fix: future annotations with custom types (#296)
Fixes #295

When combining from __future__ import annotations, Annotated, and custom types, the click parser does not have access to the custom type class at runtime unless the eval_str option is provided to inspect.signature(). Unfortunately, this option is not available for python 3.9 or below. This fixes the bug for python 3.10+ and adds a workaround in the docs for 3.9.

---------

Co-authored-by: Matthew Anderson <matt@mandersience.com>